update 09.04.2025 - реализованы дто-классы для всех сущностей - добавлены CRUD операции

+++ src/main/java/ru/pizza/eshop/repository/CartItemRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.pizza.eshop.model.CartItem;
 
 @Repository
-public interface CartItemRepository extends CrudRepository<CartItem, Integer> {
+public interface CartItemRepository extends CrudRepository<CartItem, Long> {
 }
